requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
numpy>=1.26.0
openai>=1.37.0
pydantic>=2.7.0
tiktoken>=0.7.0
langchain>=0.2.6
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
scikit-learn>=1.5.1
nltk>=3.9.1
spacy>=3.7.5
torch>=2.3.1
python-docx>=1.1.2
pyyaml>=6.0.1
jinja2>=3.1.4
